Move the backend into its own backend directory
