+++ backend/api/tests.py
@@ -1,6 +1,3 @@
-from django.test import TestCase
-from api.models import Retrospective
-
 # Create your tests here.
 if __name__ == '__main__':
     # Retrospective.create_table(read_capacity_units=1, write_capacity_units=1)
