+++ backend/retrospective/urls.py
@@ -17,6 +17,6 @@ Including another URLconf
 from django.urls import path, include
 
 urlpatterns = [
-    path('api/', include('api.urls')),
+    path('api/', include('backend.api.urls')),
     # path('admin/', admin.site.urls),
 ]
