+++ backend/tests/api/model_test.py
@@ -1,4 +1,4 @@
-from api.models import RetroStep
+from backend.api.models import RetroStep
 import pytest
